updated location of util function
