+++ app/src/components/mode-specific/desktop/InterceptTraffic/WebTraffic/TrafficTableV2/FixedRequestLogPane.js
@@ -10,7 +10,7 @@ import Editor, { EditorLanguage } from "componentsV2/CodeEditor";
 import "./FixedRequestLogPane.css";
 import { REQUEST_METHOD_COLORS, REQUEST_METHOD_BACKGROUND_COLORS } from "../../../../../../constants";
 import { RequestMethod } from "features/apiClient/types";
-import { shouldRenderPreview } from "./utils";
+import { shouldRenderPreview } from "utils/CodeEditorUtils";
 
 const { Text } = Typography;
